--- a/.gitignore
+++ b/.gitignore
@@ -1,7 +1,11 @@
 .pioenvs
-
 # python binaries, for ycm
 *.pyc
-
 # universal-ctags, for ycm/vim
 tags
+
+.clang_complete
+.gcc-flags.json
+.gch
+*.d
+*.o