--- a/.ycm_extra_conf.py
+++ b/.ycm_extra_conf.py
@@ -104,8 +104,10 @@ def MakeRelativePathsInFlagsAbsolute( flags, working_directory ):
         for path, dirs, files in os.walk(libDir):
             # Add to flags if dir contains a header file and is not
             # one of the metadata dirs (examples and extras). 
+            # don't include arduino's SD lib since we're using SdFat
             if any(IsHeaderFile(x) for x in files) and\
-              path.find("examples") is -1 and path.find("extras") is -1:
+               path.find("examples") is -1 and path.find("extras") is -1 and\
+               path.find("arduinosam/libraries/SD/src") is -1:
                 logger.debug("Directory contains header files - %s"%path)
                 flags.append('-I'+path)