--- a/platformio.ini
+++ b/platformio.ini
@@ -17,15 +17,18 @@
 # Automatic targets - enable auto-uploading
 # targets = upload
 
+[env:diecimilaatmega328]
+platform = atmelavr
+framework = arduino
+board = diecimilaatmega328
+
 [env:due]
 platform = atmelsam
 framework = arduino
 board = due
-lib_install = 1, 20, 322, 532
+lib_install = 1, 322, 532
 #[ ID  ] Name             Compatibility         "Authors": Description
 #--------------------------------------------------------------------------------------------------------------------
 #[  1  ] OneWire          arduino               "Paul Stoffregen, Jim Studt, Tom Pollard, Derek Yerger, Josh Larios, Robin James, Glenn Trewitt, Jason Dangel, Guillermo Lovato, Ken Butcher, Mark Tillotson, Bertrik Sikken, Scott Roberts": Control 1-Wire protocol (DS18S20, DS18B20, DS2408 and etc)
-#[ 20  ] Adafruit-GPS     arduino, atmelavr     "Adafruit Industries": An interrupt-based GPS library for no-parsing-required use
 #[ 322 ] SdFat            arduino, atmelavr, atmelsam, timsp430, titiva, teensy, freescalekinetis, ststm32, nordicnrf51, nxplpc, espressif, siliconlabsefm32, linux_arm, native, intel_arc32 "Bill Greiman": FAT16/FAT32 file system for SD cards
 #[ 532 ] BMP180           arduino, atmelavr     "Felix Rusu": Library for BMP180 barometric atmospheric pressure sensor, derived from Sparkfun's version with some added fixes
-