--- a/src/Globals.cpp
+++ b/src/Globals.cpp
@@ -1,0 +1,13 @@
+#include "Globals.h"
+
+void WATCHDOG_RESET() {
+	WDT_Restart(WDT);
+}
+
+// Serial streams
+ArduinoOutStream cout(Serial);
+
+// input buffer for line
+char cinBuf[40];
+ArduinoInStream cin(Serial, cinBuf, sizeof(cinBuf));
+