Big update!

Overall System
- Modules are created once on setup and globally stored in an array
- No dynamic memory allocations except in initialization
- Modules may persist data/send data over uno->radio link by overriding
the appropriate methods
- Watchdog added
- Globals.h added
  - Allow writing to Serial using c++ style cout and cin streams
  - WATCHDOG_RESET() method to appease the watchdog
- Added a project editorconfig, makefile, and improved ycm completion

Modules
- Scheduler and Tasks
  - IMPLEMENTED, WORKING: Successful multitasking of modules
- Radio
  - IMPLEMENTED, WORKING: Successful relaying of GPS telemetry
  (lat,long,time,alt) over due->uno->radio link
  - More testing needed to see if Modules can reliably relay information
  over the APRS data field
- GPS
  - IMPLEMENTED, WORKING: GPS gets satellite fixes, logs data to SD
  card, and provides RTC+telemetry to system
  - Only source of RTC for now
- SDCard
  - IMPLEMENTED, WORKING: SDCard reliably persists all data from all
  modules
- StateHandler
  - IMPLEMENTED, UNTESTED
  - Analyze GPS and Barometer values and relay state changes to system
- IMU
  - IMPLEMENTED, UNTESTED
  i2c address crashing the due on runtime)
  - Might change implementation and hardware IMU to the mpu6050
- Buzzer
  - IMPLEMENTED, UNTESTED
  - Toggles a digital pin to turn on a buzzer
- Barometer
  - IMPLEMENTED, UNTESTED
- Photocells
  - IMPLEMENTED, UNTESTED

Squashed commit of the following:

commit 98c506d5399895e2d6035827cff83b30408bb196
Author: Steven Huang <stevenhuang1995@gmail.com>
Date:   Mon Jul 25 02:56:54 2016 -0700

    Let Task constructor set own freq/interval with passed Module

    - Only calculate timestamp stuff when Task is supposed to print

commit d4a667b9020d971ce6f1564de3829aaa93b88833
Author: Steven Huang <stevenhuang1995@gmail.com>
Date:   Sun Jul 24 14:20:58 2016 -0700

    Add Buzzer module to module list, add radio dataToSend() to Module

    - Don't save Module dataToPersist() to SD if null
    - Radio module register all modules

commit 3e85cc815be12b3dbac9971eacade66038d94512
Author: Steven Huang <stevenhuang1995@gmail.com>
Date:   Sun Jul 24 10:37:04 2016 -0700

    Add project .editorconfig and autoindent sources

    - Default size-4 tabstop
    - Better serial output when printing tasks

commit d4b41ddacb740593655d9bb3a26b6d3ecf89b930
Merge: e9b8f8c 7333953
Author: Steven Huang <stevenhuang1995@gmail.com>
Date:   Sun Jul 24 05:06:13 2016 -0700

    Merge branch 'steven' of github.com:SFUSatClub/balloon1 into steven

commit e9b8f8cb7f9afcaba5b1aff13f9b6d61b3a17506
Author: Steven Huang <stevenhuang1995@gmail.com>
Date:   Sun Jul 24 04:58:06 2016 -0700

    Better serial formatting, don't tick IMU/Barometer if failed to init

    - Use snprintf to ensure no buffer overruns
    - Proper header guards for buzzer
    - GPS persist data
    - IMU triggering seg fault in due (assertion "(address & 0x80) == 0"
    failed); might need different IMU library/fix i2c addr

commit 733395341b3b6bfefff83e2e455cd36f98441662
Author: richarthurs <rich@richarthurs.com>
Date:   Sun Jul 24 02:12:48 2016 -0700

    Potential IMU crash fix

    1. enabled the IMU timeout so we don’t get stuck in a while loop.

    2. popped the magnetometer readings into the dataMagnetometer array.
    Previously, the accelerometer readings were being put in here and in
    dataAccelerometer.

    Magnetometer = compass essentially, so it’d be cool to save that data.
    GPS doesn’t give us our orientation in the same way.

commit 4b1746e04056c8814a37bbaa6549d452f882b109
Author: Steven Huang <stevenhuang1995@gmail.com>
Date:   Sat Jul 23 21:37:45 2016 -0700

    Single Wire.begin() (remove this call from barometer and imu)

    - Test Barometer and IMU with real hardware
    - Fix IMU tick() locking up system

commit 193449739b6c35d0945475ec4d849b50153be3d6
Merge: 3b69365 960f90c
Author: Steven Huang <stevenhuang1995@gmail.com>
Date:   Sat Jul 23 21:03:14 2016 -0700

    Merge pull request #12 from SFUSatClub/tim

    Globals + barometer

commit 3b69365d8cd78694a9bf964d0c12add3368a409a
Merge: f8c84ec f1411f9
Author: Steven Huang <stevenhuang1995@gmail.com>
Date:   Sat Jul 23 20:48:49 2016 -0700

    Merge branch 'steven' of github.com:SFUSatClub/balloon1 into steven

    - Add include guards into baraometer.h
    - Add missing barometer libs (heather!)

commit f8c84ecab4a33adb16bf50b970746e4626899541
Author: Steven Huang <stevenhuang1995@gmail.com>
Date:   Sat Jul 23 20:17:09 2016 -0700

    Radio 10 second interval tick

commit 960f90cb2379192568cd4b40eab6a16470af6499
Author: Tim Nosov <timnosov@gmail.com>
Date:   Sat Jul 23 19:17:21 2016 -0700

    fixed barometer merge conflict

commit ac8b58137de4a8a2005c0ef18359aa66966da434
Author: Tim Nosov <timnosov@gmail.com>
Date:   Sat Jul 23 19:16:44 2016 -0700

    included globals

commit c56b43e1427980aff77f672b754abda7b07c88ac
Author: Steven Huang <stevenhuang1995@gmail.com>
Date:   Sat Jul 23 19:12:43 2016 -0700

    Make GPS always tick (interval=0) to ensure catching TX from GPS

    - Make tasks use timeout/interval from module using
    scheduling_freq returned from Module::getSchedulingFreq()
    - Only output stuff timestamp to serial if module wants to persist data
    - Due->uno->radio bridge working

commit f1411f92fb12726669328d135163d4484e8fd566
Merge: 01d37b0 80e84b0
Author: Tim Nosov <timateus@users.noreply.github.com>
Date:   Sat Jul 23 17:33:01 2016 -0700

    Merge pull request #11 from cfinniga/steven

    Worked on Buzzer module

commit 80e84b0d7cc5398c7c20675296e94a70948cf897
Author: cfinniga <cfinnig@gmail.com>
Date:   Sat Jul 23 17:14:57 2016 -0700

    Buzzer cpp and h files have been added, but not tested yet.

commit 01d37b0d47494092e50a3b4f9a394b038d752fa3
Merge: 4c8b2db 7f900c9
Author: Lana <borklana@gmail.com>
Date:   Sat Jul 23 16:48:08 2016 -0700

    Fix main merge conflict

commit 7f900c98b50ba7eaab14499619c20a9978990e6f
Author: Lana <borklana@gmail.com>
Date:   Sat Jul 23 16:28:26 2016 -0700

    Undid change

commit 4c8b2db5bd84eff16c30983831c1223f3f1fe8d6
Merge: d9bf434 a0aa727
Author: hyyl <heather@Dagda>
Date:   Sat Jul 23 16:16:33 2016 -0700

    heather: merged barometer, statehandling, added getters to gps

commit a0aa7270c980af9c74d056339feae8b48dd53f61
Author: hyyl <heather@Dagda>
Date:   Sat Jul 23 16:03:54 2016 -0700

    merged from heather: barometer, statehandler, added getters for speed/alt to GPS

commit 8f114fca99323cb5c48810243c8909c0634fcf2f
Author: Lana <borklana@gmail.com>
Date:   Sat Jul 23 15:53:03 2016 -0700

    Added accelerometer pololu libraries

commit b761119fb90b3a83941e85365b51e55410a26e21
Author: Steven Huang <stevenhuang1995@gmail.com>
Date:   Sat Jul 23 15:28:03 2016 -0700

    Implement due->uno APRS radio bridge, GPS public interface

    - Remove task2 from scheduler
    - Fix sdcard buffer overrun :o

commit 094ee289d4cb5b0e1d88819df6eb195f5c6a5f5b
Author: hyyl <heather@Dagda>
Date:   Sat Jul 23 15:07:19 2016 -0700

    - Barometer code prepped
    - State handler prepped

    - proper values still needed for statehandler, module needs enum to be updated,main needs to change call

commit d9bf43442203dccaa4fe129b7a5ab0c47ed5f4d3
Merge: e3d9225 b8d1420
Author: Steven Huang <stevenhuang1995@gmail.com>
Date:   Sat Jul 23 14:35:39 2016 -0700

    Merge pull request #8 from cfinniga/steven

    Add platformio lib id for imu

commit b8d14206dec264a5d833ad48a6078732299e7300
Author: cfinniga <cfinnig@gmail.com>
Date:   Sat Jul 23 14:29:45 2016 -0700

    Add platformio lib id for imu

commit e3d9225828ff2282196fe503a62c44708d2f9651
Merge: cda9478 4ae8761
Author: Steven Huang <stevenhuang1995@gmail.com>
Date:   Sat Jul 23 13:54:32 2016 -0700

    Merge pull request #7 from SFUSatClub/lana

    Added IMU Module

commit 4ae8761e9a1c1319f3bc65b741ba761b5735579d
Author: Lana <borklana@gmail.com>
Date:   Sat Jul 23 13:44:12 2016 -0700

    Added IMU Module

commit 986242f599052fa2c1acf56cd2c42ad459657a06
Author: hla163 <hla163@asb9838su-d03.csil.sfu.ca>
Date:   Sat Jul 23 13:38:06 2016 -0700

    barometer basic code

commit 70952e02d3758fe47729d7c0ca809d7613fbf63f
Author: hyyl <heather@Dagda>
Date:   Sat Jul 23 12:34:58 2016 -0700

    saving because powerlow

commit fc351df1ccaf63b6048fffa6a5873279de6bc7fc
Author: hyyl <heather@Dagda>
Date:   Sat Jul 23 11:40:32 2016 -0700

    modifications to statehandler to include speed

commit cda9478186d6c0189c5e71b522be01a60e31838d
Author: Steven Huang <stevenhuang1995@gmail.com>
Date:   Thu Jul 21 21:39:29 2016 -0700

    Added watchdog stuff, scheduler is now working correctly

    - Create Globals.h
    - Fix some timing issues with scheduler; everything works now

commit 2611c9b86ee8ce818e1165a457f56b5315f8e5b9
Author: Steven Huang <stevenhuang1995@gmail.com>
Date:   Wed Jul 20 23:42:48 2016 -0700

    Explicit null check in Scheduler for stateHandler instance

    - SystemState enum use int8_t instead of int to save sram

commit dd734ca06c6d47230a95b7be1282338c5973d5d3
Author: hyyl <heather@Dagda>
Date:   Wed Jul 20 21:49:00 2016 -0700

    statehandler with presumably correct logic, still need battery code + nitpicking re: alt+pressure

commit 3abc15dd19f16920882426823e6c84a7d1466f3a
Author: Steven Huang <stevenhuang1995@gmail.com>
Date:   Wed Jul 20 20:29:20 2016 -0700

    Implement StateHandler stub and integration with Scheduler

    - Add onStatehanged() method to modules
    - Fix forgot to initialize allTasks array of pointers
    - Scheduler calls its instance of StateHandler at end of every cycle
    - If state change is detected, Scheduler propogates changes to every module

commit 5f77706e3435dae79c54614c97983ad5860310ac
Author: Steven Huang <stevenhuang1995@gmail.com>
Date:   Wed Jul 20 03:42:17 2016 -0700

    Scheduler cycles and prepare State detecting logic

commit 1dfb8de2e0d40dfdd6c05f934bcc7af480a9c0da
Author: Steven Huang <stevenhuang1995@gmail.com>
Date:   Wed Jul 20 01:06:02 2016 -0700

    Remove doxyfile from dev branches; only keep it doc branch gh-pages

commit 9cac6821ab7765ebc28b049cf6a25d7bdea2e288
Author: Steven Huang <stevenhuang1995@gmail.com>
Date:   Tue Jul 19 21:03:45 2016 -0700

    Test doxygen docs generation with Javadoc style comments

commit 2ff2222d28ca0ce564d47cada0f6cd1e0508b852
Author: Steven Huang <stevenhuang1995@gmail.com>
Date:   Tue Jul 19 20:14:11 2016 -0700

    Register all modules to be run by the Scheduler

    - Moved Scheduler.* and Task.* from libs to project src
    - Stub for module onStateChanged

commit 796bad82c048de919de5eb0ba12609d9045d0e6d
Merge: fdbd073 dfc47d6
Author: Steven Huang <stevenhuang1995@gmail.com>
Date:   Tue Jul 19 04:32:32 2016 -0700

    Merge branch 'Richard' into steven -- Untracked lib/Archive and IMU
    Functions

commit dfc47d65bd9395977a702915855628befb88e8e8
Author: Steven Huang <stevenhuang1995@gmail.com>
Date:   Tue Jul 19 04:18:41 2016 -0700

    Revert "More specific check for Due" - @richard, check refactor commit

    - Everyhting regarding ISR and tickConfig is in Scheduler.cpp

    This reverts commit 442f611992ac2713ecef9365e90717dbadb3d304.

commit fdbd073f39cbbcf89baf2385fc647353de0cace8
Author: Steven Huang <stevenhuang1995@gmail.com>
Date:   Tue Jul 19 03:45:01 2016 -0700

    Allocate photocellData dynamically, use new SDCard data format

    - Add Photocells to module list
    - Run each module's begin() via module list

commit f15ab3ae5cea0ebe20ff56bd71dd55f52beeafeb
Author: Steven Huang <stevenhuang1995@gmail.com>
Date:   Tue Jul 19 03:44:02 2016 -0700

    Fix compilations of Check* sources by commenting out IMU/GPS code for
    now

    - Added IMU stub

commit f795104780b0a1aeffb47dcd84a79990de1ac58b
Author: Steven Huang <stevenhuang1995@gmail.com>
Date:   Tue Jul 19 03:43:17 2016 -0700

    Relax pure virtual constraint of enable() and disable() when deriving Modules

    - Modules now aren't forced to implement enable()/disable() if there's no reason to

commit 442f611992ac2713ecef9365e90717dbadb3d304
Author: richarthurs <Rich@Richards-Mac-Pro.local>
Date:   Mon Jul 18 19:53:30 2016 -0700

    More specific check for Due

commit f5cdac53a495e2d8e3402d603dea580048047e52
Author: Steven Huang <stevenhuang1995@gmail.com>
Date:   Mon Jul 18 02:36:22 2016 -0700

    Vim users only: don't tell clang to include arduino's implementation of sdfat else it gets confused

commit bff6983e62513ee53c27bea88f72eb4e6bc390cd
Merge: be1f4ed 4dadd2d
Author: Steven Huang <stevenhuang1995@gmail.com>
Date:   Mon Jul 18 01:32:22 2016 -0700

    Merge branch 'heather' into steven

commit be1f4ed062a5a3a4a73030051a00ec3c2735ec4c
Author: Steven Huang <stevenhuang1995@gmail.com>
Date:   Mon Jul 18 01:23:44 2016 -0700

    Generate docs using doxygen and graphviz

commit de3c71af08c99d496ad36c52286deae3059c3009
Author: Steven Huang <stevenhuang1995@gmail.com>
Date:   Sun Jul 17 15:27:33 2016 -0700

    Add comments and split up big boolean expression

    - Set and use the scheduler's systemTick

commit 296556d05332f11a1a358e9a4e605324f11bb79b
Author: Steven Huang <stevenhuang1995@gmail.com>
Date:   Sun Jul 17 15:02:30 2016 -0700

    Refactor into Tasks and Scheduler classes

    Untested code! Richard, please check if works in runtime of both due and
    atmega

    - Moved all ISR and flag stuff into Scheduler
    - Add gch to gitignore
    - Let scheduler manage running of tasks
    - Warning: current implementation of scheduler tick uses shared static
    variable; don't try to use more than one scheduler
    - System tick passed as argument into Task::runTask()
    - Task::runTask() sets their own lastRun instead of scheduler

commit bdab282b2b4dae6a947815b66a5a66cb01367566
Merge: 438cb8d 572a348
Author: richarthurs <Rich@Richards-Mac-Pro.local>
Date:   Sun Jul 17 01:29:19 2016 -0700

    Merge branch 'RichardDue' into Richard

commit 572a348cee44466e9fb46b81d5b4aa57ed7f9e65
Author: richarthurs <Rich@Richards-Mac-Pro.local>
Date:   Sun Jul 17 01:28:22 2016 -0700

    Documented some more things

    sorry for all the commits…

commit b301e65fee7c4aba90d958e9410cf020a9131b5a
Author: richarthurs <Rich@Richards-Mac-Pro.local>
Date:   Sun Jul 17 01:25:16 2016 -0700

    Cleaned up

    Update: can confirm the timing is really really accurate.
    Also works on the 328 without changing anything so that’s solid too.

commit 097404354dc95e376a403bf6a8358920fbeb964b
Author: richarthurs <Rich@Richards-Mac-Pro.local>
Date:   Sun Jul 17 01:17:17 2016 -0700

    Working multitasking on Due

    Multitasking works on the Due
    Should work on 328 and Due because of the processor checks

    Todo: add watchdog for the Due, figure out a nicer way to deal with the
    scheduler table, integrate some fancier things than just blinking LEDs.

commit 438cb8dd6487c20ebd6858dcab61650b0cf73879
Author: richarthurs <Rich@Richards-Mac-Pro.local>
Date:   Sat Jul 16 23:47:42 2016 -0700

    Working 328 branch

commit 4e4cb062cb8f68683858a1b3bcb5ab9033e65c4a
Author: Steven Huang <stevenhuang1995@gmail.com>
Date:   Sat Jul 16 23:28:27 2016 -0700

    fix merge

commit 703511352ca2b168309c1db4e2a47bba8aa4504a
Merge: 494bc56 35faa3e
Author: richarthurs <Rich@Richards-Mac-Pro.local>
Date:   Sat Jul 16 23:21:29 2016 -0700

    Merge remote-tracking branch 'origin/Richard' into Richard

    Conflicts:
    	src/main.cpp

commit 494bc5643a725e72c3bf572df2d1a392c1fd94ac
Author: richarthurs <Rich@Richards-Mac-Pro.local>
Date:   Sat Jul 16 23:20:19 2016 -0700

    Removed watchdog for 328

    The duemilanove board I’m using takes long enough to bring up the boot
    loader that the watchdog fires. So it just blinks pin 13 once the
    watchdog is hit the first time. Can fix with optiboot, but whatever…
    https://bigdanzblog.wordpress.com/2014/10/23/installing-the-optiboot-loa
    der-on-an-arudino-nano-to-fix-the-watch-dog-timer-wdt-issue/

commit 35faa3e55667d21bb9e07ba0db92e63e8769ab8e
Author: Steven Huang <stevenhuang1995@gmail.com>
Date:   Sat Jul 16 23:08:20 2016 -0700

    Neaten things up

    - Fill up allTasks array directly using dynamic allocation since they'll
    end up in heap anyways
    - Make TaskIndex a local var to faciliate portability

commit af591686c66c20dcb0a26861a91384e466edc05b
Author: richarthurs <Rich@Richards-Mac-Pro.local>
Date:   Sat Jul 16 21:30:00 2016 -0700

    Prepped for watchdog test

commit c67730e940d8a44e980bd662a27384c51842748b
Author: richarthurs <rich@richarthurs.com>
Date:   Sat Jul 16 20:50:31 2016 -0700

    Working on timeout

commit af6000de9c73069f3a3470d9a8386d8cd5dae587
Author: Steven Huang <stevenhuang1995@gmail.com>
Date:   Sat Jul 16 19:45:48 2016 -0700

    Remove unneeded delay

commit bb7dd451b68cf423d2c2db836993d5ff2e50b239
Author: richarthurs <rich@richarthurs.com>
Date:   Sat Jul 16 18:21:38 2016 -0700

    Added check for multiple runs.

    Safety feature, ensures tasks will only run once if they happen to be
    called between ticks.

commit 5de8f18d3ba4c17bc26d7c8110bb5487bf2de08e
Author: richarthurs <rich@richarthurs.com>
Date:   Sat Jul 16 18:15:00 2016 -0700

    Scheduling works :)

    Steven sorted out the extern business, now have access to the system
    clock in Task.run

commit b164d0eaebca58e0e8603c4b020294d86fe650a5
Author: richarthurs <rich@richarthurs.com>
Date:   Sat Jul 16 16:48:24 2016 -0700

    Cleaned up, working tasks.

    Timing has been confirmed to be pretty accurate and stable over minute
    timeframes.

commit 36fa78a218b86da27fe6f6915c042c93f058f7cd
Author: richarthurs <rich@richarthurs.com>
Date:   Sat Jul 16 15:22:49 2016 -0700

    Pointer to tasks for stability

commit fff09a82b99083de3f8dc042a2dcf36b609feb33
Author: richarthurs <rich@richarthurs.com>
Date:   Sat Jul 16 14:31:53 2016 -0700

    Added function to run to constructor

    Works but with weird timing issues. Will switch to pointer method used
    before.

commit f194011ef703afeb6ab3ed06d40b35bb19521996
Author: richarthurs <rich@richarthurs.com>
Date:   Sat Jul 16 12:44:04 2016 -0700

    Revert "Sort of working array of objects"

    This reverts commit 2ab6b6023e33a54c764df05f811e6fe1b7bce7bd.

commit 2ab6b6023e33a54c764df05f811e6fe1b7bce7bd
Author: richarthurs <Rich@Richards-Mac-Pro.local>
Date:   Sat Jul 16 12:31:06 2016 -0700

    Sort of working array of objects

    … timing errors are back :(

commit ca2a2022d9c48426032746a114b86912976a48fd
Author: richarthurs <Rich@Richards-Mac-Pro.local>
Date:   Sat Jul 16 12:08:00 2016 -0700

    Fixed instability - new working commit

    Tested to run stable for about an hour. Issue was trying to store a
    32-bit number in a 16-bit variable, added check to only run tasks once
    within a tick.

commit a258c911c66c6508d362f045094a7bd5fc8448ab
Author: richarthurs <Rich@Richards-Mac-Pro.local>
Date:   Sat Jul 16 11:11:20 2016 -0700

    Revert ec98654..2e71bd9

    This rolls back to commit ec9865419814c01374897ed76cebce304d37af75.

commit 2e71bd9e04bb870b2ba5e5cbba7759ec84cfb976
Author: richarthurs <Rich@Richards-Mac-Pro.local>
Date:   Sat Jul 16 11:11:07 2016 -0700

    broke it

commit ec9865419814c01374897ed76cebce304d37af75
Author: richarthurs <Rich@Richards-Mac-Pro.local>
Date:   Fri Jul 15 01:47:58 2016 -0700

    Sort of working task class

    Stops after a while and doesn’t have multiple tasks going. But the idea
…